fix(TD-036): consolidate to_serializable into routes/_shared.py
api/sse.py's to_serializable duplicated routes/execute.py's private
_to_serializable byte-for-byte (only the name and docstring differed).
Moved the single implementation into routes/_shared.py — the existing
home for cross-route helpers alongside normalize_graph_identifier —
and updated both execute.py and sse.py to import it instead of
maintaining separate copies.

Research (TD-036.research-report.md) also found that the originally
reported _normalize_graph_identifier duplication in stream.py did not
actually exist: both execute.py and stream.py already import the
canonical normalize_graph_identifier from _shared.py, so no change was
needed there.

Zero behavior change: verified execute.to_serializable is
sse.to_serializable is _shared.to_serializable (same function object).
No import cycle (module-level import; _shared.py has no dependency on
execute.py/sse.py). make lint clean; targeted SSE + execute test
suites (test_sse_stream_endpoint, test_sse_execute_streaming,
test_sse_config, test_execution_routes_fixed, test_async_route_facade,
test_http_execute_suspend_resume) all pass.
